+++ amazon_security_lake/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 1.1.0
+## Added
+- Added multiple authentication methods.
+
+## Changed
+- Updated Mapping file.
+
+## Fixed
+- Bug fix.
+
+
+# 1.0.1
+## Changed
+- Changed structure for unmapped data while pushing the data to Amazon Security Lake.
+
+
 # 1.0.0
 ## Added
 - Initial release.